♻️ [Refactoring]: #432 [FE] BoardWorkspace の ActiveBoardView / BoardView を独立ファイルへ抽出
BoardWorkspace に同居していた viewMode 分岐器 ActiveBoardView と board 分岐ブロックを
それぞれ独立フォルダ（ActiveBoardView/ ・ BoardView/）へ抽出し BoardWorkspace をスリム化。
挙動不変の純粋リファクタリングで、参照用に BoardWorkspaceProps を export 化する。

Refs #432
